requirements.txt
# requirements.txt - Install with: pip install -r requirements.txt

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP requests
requests==2.31.0
aiohttp==3.9.1

# Google Gemini AI
google-generativeai==0.3.2

# Data validation
pydantic==2.5.0

# CORS support
python-multipart==0.0.6

# Development dependencies (optional)
python-dotenv==1.0.0  # For environment variables if you want to use .env files later